pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aiz"
version = "0.0.1" # Start with 0.0.1 for a placeholder
authors = [
  { name="polymorphisma", email="shrawansunar.6@gmail.com" },
]
description = "AI-Powered CLI Navigator. Reserved for future development."
readme = "README.md"
requires-python = ">=3.9.0, <4.0" # Let's stick with 3.8+ as planned
license = { text = "MIT License" } # Good practice to declare a license
keywords = ["ai", "cli", "shell", "developer-tools", "llm"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Environment :: Console",
    "Topic :: Utilities",
]
dependencies = [
    "langchain>=0.0.27",
    "langchain-anthropic>=0.1.23",
    "langchain-aws>=0.1.18",
    "langgraph>=0.6.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.4",
]

[project.urls]
Homepage = "https://github.com/polymorphisma/aiz" # Add your GitHub repo URL later
Repository = "https://github.com/polymorphisma/aiz"
